+++ changelog.md
@@ -19,6 +19,7 @@ Check the command for more infos.
 * Shulker bullets now take duration and amplifier from owners so each shulker can have its own
 * Snow golems can now strafe when attacking (disabled by default)
 * Fixed temporary cobwebs breaking any block in the cobweb's position
+* Fixed disabling "Melee Attacking" feature making melee attacks have no cooldown
 
 ## 2.7.1
 * Villagers no longer use target changes